osx/ios build fixes

+++ foundation/memory.c
@@ -94,7 +94,7 @@ static void* _memory_allocate_malloc( uint64_t size, unsigned int align, memory_
 	void* memory = 0;
 	if( !align )
 		return malloc( (size_t)size );
-	int result = posix_memalign( &memory, align, size );
+	int result = posix_memalign( &memory, align, (size_t)size );
 	if( result || !memory )
 		log_errorf( ERRORLEVEL_PANIC, ERROR_OUT_OF_MEMORY, "Unable to allocate memory: %s", system_error_message( 0 ) );
 	return ( result == 0 ) ? memory : 0;
@@ -123,11 +123,11 @@ static void* _memory_reallocate_malloc( void* p, uint64_t size, unsigned int ali
 #else
 	if( align )
 	{
-		void* memory = realloc( p, size + align );
+		void* memory = realloc( p, (size_t)( size + align ) );
 		memory = _memory_align_pointer( memory, align );
 		return memory;
 	}
-	return realloc( p, size );
+	return realloc( p, (size_t)size );
 #endif
 }
